Committing wip to change docIsSyncing to take into account the schema. This doesn't work because of a dependency cycle, but I wanted to commit it.
